lstopo: fix some cast warnings
Thanks to Barry M. Tannenbaum.

+++ hwloc/bitmap.c
@@ -248,7 +248,7 @@ int hwloc_bitmap_snprintf(char * __hwloc_restrict buf, size_t buflen, const stru
       return -1;
     ret += res;
     if (res >= size)
-      res = size>0 ? size - 1 : 0;
+      res = size>0 ? (int)size - 1 : 0;
     tmp += res;
     size -= res;
   }
@@ -298,7 +298,7 @@ int hwloc_bitmap_snprintf(char * __hwloc_restrict buf, size_t buflen, const stru
 #endif
 
     if (res >= size)
-      res = size>0 ? size - 1 : 0;
+      res = size>0 ? (int)size - 1 : 0;
 
     tmp += res;
     size -= res;
@@ -429,7 +429,7 @@ int hwloc_bitmap_list_snprintf(char * __hwloc_restrict buf, size_t buflen, const
     ret += res;
 
     if (res >= size)
-      res = size>0 ? size - 1 : 0;
+      res = size>0 ? (int)size - 1 : 0;
 
     tmp += res;
     size -= res;
@@ -533,7 +533,7 @@ int hwloc_bitmap_taskset_snprintf(char * __hwloc_restrict buf, size_t buflen, co
       return -1;
     ret += res;
     if (res >= size)
-      res = size>0 ? size - 1 : 0;
+      res = size>0 ? (int)size - 1 : 0;
     tmp += res;
     size -= res;
   }
@@ -569,7 +569,7 @@ int hwloc_bitmap_taskset_snprintf(char * __hwloc_restrict buf, size_t buflen, co
       return -1;
     ret += res;
     if (res >= size)
-      res = size>0 ? size - 1 : 0;
+      res = size>0 ? (int)size - 1 : 0;
     tmp += res;
     size -= res;
   }
@@ -627,7 +627,7 @@ int hwloc_bitmap_taskset_sscanf(struct hwloc_bitmap_s *set, const char * __hwloc
   }
   /* we know there are other characters now */
 
-  chars = strlen(current);
+  chars = (int)strlen(current);
   count = (chars * 4 + HWLOC_BITS_PER_LONG - 1) / HWLOC_BITS_PER_LONG;
 
   hwloc_bitmap_reset_by_ulongs(set, count);
